refactor(components/form): move inputs into the base directory
